auth_service/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-dotenv